fraud-detection/pyproject.toml
[project]
name = "fraud-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "zurdavid", email = "david.z@sunrise.ch" }
]
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "polars>=1.30.0",
    "pyarrow>=20.0.0",
    "scikit-learn>=1.6.1",
    "torch>=2.7.0",
    "xgboost-cpu>=3.0.2",
]

[tool.ruff.lint]
select = ["E", "F", "B", "I", "RUF", "PL"]  

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
